Support for ListenerAggregates in SharedEventManager

+++ library/Zend/EventManager/SharedEventManager.php
@@ -56,13 +56,27 @@ class SharedEventManager implements SharedEventManagerInterface
      * </code>
      *
      * @param  string|array $id Identifier(s) for event emitting component(s)
-     * @param  string $event
-     * @param  callable $callback PHP Callback
+     * @param  string|ListenerAggregateInterface $event An event name. If a ListenerAggregateInterface, proxies to {@link attachAggregate()}.
+     * @param  callable|int $callback If string $event provided, expects PHP callback; for a ListenerAggregateInterface $event, this will be the priority
      * @param  int $priority Priority at which listener should execute
-     * @return CallbackHandler|array Either CallbackHandler or array of CallbackHandlers
+     * @return CallbackHandler|array Either CallbackHandler or array of CallbackHandlers; mixed if attaching aggregate
+     * @throws Exception\InvalidArgumentException
      */
-    public function attach($id, $event, $callback, $priority = 1)
+    public function attach($id, $event, $callback = null, $priority = 1)
     {
+        // Proxy ListenerAggregateInterface arguments to attachAggregate()
+        if ($event instanceof ListenerAggregateInterface) {
+            return $this->attachAggregate($id, $event, $callback);
+        }
+
+        // Null callback is invalid
+        if (null === $callback) {
+            throw new Exception\InvalidArgumentException(sprintf(
+                    '%s: expects a callback; none provided',
+                    __METHOD__
+            ));
+        }
+
         $ids = (array) $id;
         $listeners = array();
         foreach ($ids as $id) {
@@ -78,18 +92,78 @@ class SharedEventManager implements SharedEventManagerInterface
     }
 
     /**
+     * Attach a listener aggregate
+     *
+     * Listener aggregates accept an EventManagerInterface instance, and call attach()
+     * one or more times, typically to attach to multiple events using local
+     * methods.
+     *
+     * @param  string $id Identifier for event emitting component
+     * @param  ListenerAggregateInterface $aggregate
+     * @param  int $priority If provided, a suggested priority for the aggregate to use
+     * @return mixed return value of {@link ListenerAggregateInterface::attach()}
+     */
+    public function attachAggregate($id, ListenerAggregateInterface $aggregate, $priority = 1)
+    {
+        $ids = (array) $id;
+        $results = array();
+        foreach ($ids as $id) {
+            if (!array_key_exists($id, $this->identifiers)) {
+                $this->identifiers[$id] = new EventManager();
+            }
+            $results[] = $aggregate->attach($this->identifiers[$id], $priority);
+        }
+        if (count($results) > 1) {
+            return $results;
+        }
+        return $results[0];
+    }
+
+
+    /**
      * Detach a listener from an event offered by a given resource
      *
-     * @param  string|int $id
-     * @param  CallbackHandler $listener
+     * @param  string|int $id Identifier for event emitting component
+     * @param  CallbackHandler|ListenerAggregateInterface $listener
      * @return bool Returns true if event and listener found, and unsubscribed; returns false if either event or listener not found
+     * @throws Exception\InvalidArgumentException if invalid listener provided
      */
-    public function detach($id, CallbackHandler $listener)
+    public function detach($id, $listener)
     {
+        if ($listener instanceof ListenerAggregateInterface) {
+            return $this->detachAggregate($id, $listener);
+        }
+
+        if (!$listener instanceof CallbackHandler) {
+            throw new Exception\InvalidArgumentException(sprintf(
+                    '%s: expected a ListenerAggregateInterface or CallbackHandler; received "%s"',
+                    __METHOD__,
+                    (is_object($listener) ? get_class($listener) : gettype($listener))
+            ));
+        }
+
         if (!array_key_exists($id, $this->identifiers)) {
             return false;
         }
         return $this->identifiers[$id]->detach($listener);
+    }
+
+    /**
+     * Detach a listener aggregate
+     *
+     * Listener aggregates accept an EventManagerInterface instance, and call detach()
+     * of all previously attached listeners.
+     *
+     * @param  string|int $id Identifier for event emitting component
+     * @param  ListenerAggregateInterface $aggregate
+     * @return mixed return value of {@link ListenerAggregateInterface::detach()}
+     */
+    public function detachAggregate($id, ListenerAggregateInterface $aggregate)
+    {
+        if (!array_key_exists($id, $this->identifiers)) {
+            return false;
+        }
+        return $aggregate->detach($this->identifiers[$id]);
     }
 
     /**
